modification de type

+++ BackEnd_GestionEvennement/src/main/java/com/groupe5/domaine/ServiceRepository.java
@@ -2,6 +2,6 @@ package com.groupe5.domaine;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ServiceRepository extends CrudRepository<Service, Integer>{
+public interface ServiceRepository extends CrudRepository<Service, Long>{
     
 }
